Move app options to own submodule
